back-end/tasks/requirements.txt
asgiref==3.11.0
Django==6.0.1
djangorestframework==3.16.1
drf-yasg==1.21.14
inflection==0.5.1
iniconfig==2.3.0
Markdown==3.10
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==9.0.2
pytest-django==4.11.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
